# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Astral Core v7 API
  version: 7.0.0
  description: |
    Comprehensive mental health platform API providing secure endpoints for:
    - User authentication and MFA management
    - Crisis assessment and intervention
    - Wellness data tracking and analytics
    - Payment processing and subscription management
    - Appointment scheduling and session management
    - Secure messaging and file management
    
    **Security Features:**
    - Rate limiting on all endpoints
    - PHI-compliant data encryption
    - Comprehensive audit logging
    - Role-based access control
    
    **Base URL:** `https://astral-core.app/api`
    
  contact:
    name: Astral Core API Support
    email: api-support@astral-core.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://astral-core.app/api
    description: Production server
  - url: https://staging.astral-core.app/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
security:
  - bearerAuth: []
  - sessionAuth: []
paths:
...
  /treatment-plans:
    get:
      tags: [Treatment Plans]
      summary: Get treatment plans
      description: Retrieve treatment plans (clients see their own, therapists see assigned plans)
      operationId: getTreatmentPlans
      parameters:
        - name: status
          in: query
          description: Filter by plan status
          schema:
            type: string
            enum: [DRAFT, ACTIVE, COMPLETED, DISCONTINUED]
        - name: clientId
          in: query
          description: Filter by client ID (therapists only)
          schema:
            type: string
      responses:
        '200':
          description: Treatment plans retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TreatmentPlan'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Treatment Plans]
      summary: Create treatment plan
      description: Create a new treatment plan (therapists only)
      operationId: createTreatmentPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTreatmentPlanRequest'
      responses:
        '201':
          description: Treatment plan created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TreatmentPlan'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Therapist access required
        '500':
          $ref: '#/components/responses/ServerError'
...
  /therapist/session-notes:
    get:
      tags: [Therapist, Session Notes]
      summary: Get session notes
      description: Retrieve session notes (therapists only)
      operationId: getSessionNotes
      parameters:
        - name: clientId
          in: query
          description: Filter by client ID
          schema:
            type: string
        - name: appointmentId
          in: query
          description: Filter by appointment ID
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Session notes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionNote'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Therapist access required
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Therapist, Session Notes]
      summary: Create session note
      description: Create a new session note (therapists only)
      operationId: createSessionNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionNoteRequest'
      responses:
        '201':
          description: Session note created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/SessionNote'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Therapist access required
        '500':
          $ref: '#/components/responses/ServerError'

  # ==================== NOTIFICATIONS PREFERENCES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
...
    TreatmentPlan:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        therapistId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [DRAFT, ACTIVE, COMPLETED, DISCONTINUED]
        goals:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              description:
                type: string
              targetDate:
                type: string
                format: date
              objectives:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    description:
                      type: string
                    isCompleted:
                      type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTreatmentPlanRequest:
      type: object
      required:
        - clientId
        - title
        - description
        - goals
      properties:
        clientId:
          type: string
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        goals:
          type: array
          items:
            type: object
            required:
              - title
              - description
            properties:
              title:
                type: string
              description:
                type: string
              targetDate:
                type: string
                format: date
              objectives:
                type: array
                items:
                  type: object
                  required:
                    - description
                  properties:
                    description:
                      type: string
...
    SessionNote:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        therapistId:
          type: string
        appointmentId:
          type: string
        sessionDate:
          type: string
          format: date-time
        notes:
          type: string
        assessments:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              score:
                type: number
              notes:
                type: string
        interventions:
          type: array
          items:
            type: string
        goals:
          type: array
          items:
            type: string
        nextSteps:
          type: array
          items:
            type: string
        isConfidential:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateSessionNoteRequest:
      type: object
      required:
        - clientId
        - appointmentId
        - notes
      properties:
        clientId:
          type: string
        appointmentId:
          type: string
        sessionDate:
          type: string
          format: date-time
        notes:
          type: string
          maxLength: 5000
        assessments:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              score:
                type: number
              notes:
                type: string
        interventions:
          type: array
          items:
            type: string
        goals:
          type: array
          items:
            type: string
        nextSteps:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "An error occurred"
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              path:
                type: array
                items:
                  type: string
          description: Detailed validation errors (when applicable)

  # ==================== COMMON RESPONSES ====================
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: Unauthorized access
              value:
                error: "Authentication required"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_failed:
              summary: Validation failed
              value:
                error: "Validation failed"
                details:
                  - code: "invalid_type"
                    message: "Expected string, received number"
                    path: ["email"]
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Server error
              value:
                error: "An internal server error occurred"
...